# docs/api-contract.openapi.yaml
openapi: 3.0.3
info:
  title: F-Solution API
  version: "1.0.0"
  description: >
    Hợp đồng REST API dùng chung cho apps/backend (NestJS) và apps/frontend (React).
    Đây là nguồn sự thật contract-first — mọi endpoint mới phải được thêm ở đây
    trước khi Claude Code CLI viết code trong /build.
servers:
  - url: /api/v1
paths:
...
  /orders:
    get:
      operationId: getOrders
      summary: >
        Lấy danh sách hợp đồng (response nhúng kèm items[] — mới Phase 3, tránh phải gọi thêm
        GET /orders/{id}/items cho từng dòng). Mới (Phase 3): nếu tài khoản đăng nhập link tới 1
        Employee role=Sale (không phải admin), tự động chỉ trả về Order có assignedTo=chính Sale đó.
      tags: [Sales]
      parameters:
        - name: assignedTo
          in: query
          required: false
          schema: { type: string }
          description: >
            Mới (Phase 3) — truyền 'me' để lọc theo chính Sale đang đăng nhập, hoặc truyền thẳng
            employeeId. LƯU Ý bảo mật: nếu tài khoản gọi API là Sale bị giới hạn (không phải admin),
            giá trị này bị BỎ QUA hoàn toàn — server luôn ép về đúng chính họ, không thể ghi đè để
            xem Order của Sale khác. Chỉ có tác dụng lọc tự do với admin/tài khoản không bị giới hạn.
      responses:
        "200":
          description: Lấy danh sách thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      summary: >
        Tạo hợp đồng mới. Mới (Phase 3): nhận trực tiếp thông tin khách hàng + 1 sản phẩm (thay vì
        chỉ customerId) — Service tự upsert Customer theo shippingPhone (chưa tồn tại → tạo mới
        customerType=new; đã tồn tại → cập nhật customerType=old) và tạo kèm 1 OrderItem, cùng
        1 $transaction. customerId (cũ) vẫn nhận được để tương thích ngược, nhưng FE nên dùng bộ
        field mới bên dưới.
      tags: [Sales]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Tạo thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: >
            Thiếu customerId lẫn customerName+shippingPhone; hoặc leadId không tồn tại/chưa
            status=converted/không thuộc chính Sale đang gọi API (mới Phase 3)
  /orders/{id}:
    patch:
      operationId: updateOrder
      summary: Cập nhật trạng thái hợp đồng (mới — Phase 2; khoá sửa sau khi status=delivered theo nguyên tắc #6)
      tags: [Sales]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderInput" }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/Order" } } }
        "409": { description: "Đơn đã ở trạng thái khoá (delivered/cancelled), không cho sửa" }
  /orders/{id}/items:
    get:
      operationId: getOrderItems
      summary: Danh sách sản phẩm/quà tặng trong 1 hợp đồng (mới — Phase 2, giải quyết bài toán combo/quà tặng)
      tags: [Sales]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/OrderItem" } } } }
    post:
      operationId: addOrderItem
      summary: Thêm sản phẩm/quà tặng vào hợp đồng
      tags: [Sales]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderItemInput" }
      responses:
        "201":
          description: Tạo thành công
          content: { application/json: { schema: { $ref: "#/components/schemas/OrderItem" } } }
...
  /accounts:
    get:
      operationId: listAccounts
      summary: Danh sách tài khoản
      tags: [account-management]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AccountProfile" } }
    post:
      operationId: createAccount
      summary: Tạo tài khoản mới
      tags: [account-management]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAccountInput" }
      responses:
        "201":
          description: Tạo thành công
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountProfile" }
  /accounts/{id}:
    patch:
      operationId: updateAccount
      summary: Sửa tài khoản (đổi tên hiển thị, reset mật khẩu)
      tags: [account-management]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAccountInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountProfile" }
    delete:
      operationId: deleteAccount
      summary: Xoá tài khoản
      tags: [account-management]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "204": { description: Đã xoá }
        "409": { description: Không thể xoá tài khoản admin cuối cùng }
  /accounts/{id}/permissions:
    get:
      operationId: getAccountPermissions
      summary: Xem ma trận quyền của 1 tài khoản
      tags: [account-management]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Permission" } }
    put:
      operationId: setAccountPermissions
      summary: Cập nhật ma trận quyền của 1 tài khoản
      tags: [account-management]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: array, items: { $ref: "#/components/schemas/PermissionInput" } }
      responses:
        "200": { description: OK }

  # ============================================================
  # Phase 2 — MỚI, theo MÔ TẢ NGHIỆP VỤ CỤ THỂ.pdf (V3.0). Xem
  # docs/implementation_plan.md mục "Danh sách entity mới (Phase 2)" cho lý do
  # thiết kế và bảng ánh xạ module permission.
  # ============================================================
...
  /marketing-reports:
    get:
      operationId: getMarketingReports
      summary: Báo cáo hiệu suất Marketing nhập tay theo ngày/sản phẩm/thị trường (dữ liệu báo cáo, không phải nguồn sự thật)
      tags: [MKT]
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/MarketingReport" } } } }
    post:
      operationId: createMarketingReport
      summary: Nhập báo cáo Marketing
      tags: [MKT]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateMarketingReportInput" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/MarketingReport" } } }
  /marketing-reports/auto:
    get:
      operationId: getMarketingReportsAuto
      summary: >
        Mới — Báo cáo tự động Marketing, tính real-time (KHÔNG lưu bảng riêng) từ Lead.sourcedBy +
        Order theo khoảng ngày. Doanh thu map giữa nhân viên Marketing và doanh thu các Lead họ tạo.
      tags: [MKT]
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
        - name: toDate
          in: query
          required: false
          description: "Gộp số liệu theo khoảng [date, toDate]. Bỏ trống thì chỉ tính 1 ngày (= date)."
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/MarketingReportAuto" } } } }
        "400": { description: Thiếu query param 'date' }
  /customers:
    get:
      operationId: getCustomers
      summary: Lấy danh sách khách hàng (đã chuẩn hoá, tránh trùng lặp)
      tags: [Sales]
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Customer" } } } }
    post:
      operationId: createCustomer
      summary: Tạo khách hàng mới
      tags: [Sales]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateCustomerInput" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/Customer" } } }
  /customers/{id}:
    patch:
      operationId: updateCustomer
      summary: Sửa khách hàng (địa chỉ, loại khách, blacklist)
      tags: [Sales]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateCustomerInput" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Customer" } } } }
...
  /payments:
    get:
      operationId: getPayments
      summary: Thu tiền theo hợp đồng, đối soát bill
      tags: [Finance]
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Payment" } } } }
    post:
      operationId: createPayment
      summary: Ghi nhận khoản thu cho 1 hợp đồng
      tags: [Finance]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreatePaymentInput" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/Payment" } } }
  /payments/{id}:
    patch:
      operationId: updatePayment
      summary: Cập nhật trạng thái/đối soát khoản thu (không có bill thì không được set completed, theo nguyên tắc #3)
      tags: [Finance]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdatePaymentInput" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Payment" } } } }
...
  /employees:
    get:
      operationId: getEmployees
      summary: Hồ sơ nhân sự (nghiệp vụ — khác Account đăng nhập, xem implementation_plan.md)
      tags: [HR]
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Employee" } } } }
    post:
      operationId: createEmployee
      summary: Tạo hồ sơ nhân sự mới
      tags: [HR]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateEmployeeInput" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/Employee" } } }
  /employees/{id}:
    patch:
      operationId: updateEmployee
      summary: Sửa hồ sơ nhân sự (team, branch, role, trạng thái, liên kết Account)
      tags: [HR]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateEmployeeInput" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Employee" } } } }
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, customerId, totalAmount, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          description: ID khách hàng đặt đơn — tham chiếu Customer (xem /customers, mới ở Phase 2)
        customerName:
          type: string
          nullable: true
          description: "Mới — sao chép từ Customer.name tại thời điểm GET /orders, để FE hiển thị được kể cả khi tài khoản không có quyền GET /customers"
        leadId:
          type: string
          format: uuid
          nullable: true
          description: "Mới (Phase 3) — Lead phát sinh đơn này (nếu có), tham chiếu lỏng Lead.id"
        leadCode:
          type: string
          nullable: true
          description: "Mới — sao chép từ Lead.code tại thời điểm GET /orders, dùng để thống kê doanh thu theo Lead"
        assignedTo:
          type: string
          format: uuid
          nullable: true
          description: "Mới (Phase 3) — Sale phụ trách đơn này, tham chiếu Employee"
        assignedToName:
          type: string
          nullable: true
          description: "Mới — sao chép từ Employee.name tại thời điểm GET /orders, để FE hiển thị được kể cả khi tài khoản không có quyền GET /employees"
        shippingAddress: { type: string, nullable: true, description: "Mới (Phase 3) — Địa chỉ nhận hàng" }
        shippingPhone: { type: string, nullable: true, description: "Mới (Phase 3) — SĐT nhận hàng" }
        note: { type: string, nullable: true, description: "Mới (Phase 3) — Ghi chú hợp đồng" }
        totalAmount:
          type: number
          format: double
          description: Tổng giá trị hợp đồng (VNĐ)
        status:
          type: string
          description: Trạng thái xử lý hợp đồng
          enum: [pending, confirmed, shipped, delivered, cancelled]
        items:
          type: array
          description: "Mới (Phase 3) — nhúng kèm trong response GET /orders để hiển thị đủ cột Sản phẩm/Số lượng/Giá bán/%KM"
          items: { $ref: "#/components/schemas/OrderItem" }
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrderInput:
      type: object
      required: [totalAmount]
      description: >
        Mới (Phase 3): customerId trở thành optional (giữ tương thích ngược) — cách tạo đơn khuyến
        nghị là truyền customerName + shippingPhone (+ shippingAddress) để Service tự upsert Customer,
        cùng thông tin sản phẩm để tạo kèm 1 OrderItem đầu tiên.
      properties:
        customerId:
          type: string
          minLength: 1
          description: "Cũ (Phase 1) — vẫn nhận, nhưng ưu tiên dùng customerName/shippingPhone bên dưới"
        leadId: { type: string, format: uuid, description: "Mới (Phase 3) — chỉ nhận Lead có status=converted và assignedTo=chính Sale đang tạo đơn" }
        assignedTo: { type: string, format: uuid, description: "Mới (Phase 3) — Sale phụ trách; mặc định = chính tài khoản Sale đang gọi API nếu không truyền" }
        customerName: { type: string, minLength: 1, description: "Mới (Phase 3) — dùng để upsert Customer theo shippingPhone" }
        shippingAddress: { type: string, description: "Mới (Phase 3)" }
        shippingPhone: { type: string, description: "Mới (Phase 3) — khoá upsert Customer (Customer.phone unique)" }
        note: { type: string, description: "Mới (Phase 3)" }
        productId: { type: string, format: uuid, description: "Mới (Phase 3) — tạo kèm 1 OrderItem đầu tiên" }
        productName: { type: string, description: "Mới (Phase 3)" }
        unit: { type: string, description: "Đơn vị tính — sao chép từ Product đã chọn" }
        quantity: { type: integer, minimum: 1, description: "Mới (Phase 3)" }
        unitPrice: { type: number, minimum: 0, description: "Mới (Phase 3)" }
        discountPercent: { type: number, minimum: 0, maximum: 100, description: "Mới (Phase 3) — % Khuyến mại" }
        totalAmount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          description: "Thành tiền — FE tự tính quantity*unitPrice*(1-discountPercent/100), Service không tính lại"
    UpdateOrderInput:
      type: object
      description: Mới (Phase 2) — chặn sửa khi status hiện tại đã là delivered/cancelled (nguyên tắc #6)
      properties:
        status: { type: string, enum: [pending, confirmed, shipped, delivered, cancelled] }
    OrderItem:
      type: object
      description: Mới (Phase 2) — giải quyết bài toán gói combo nhiều sản phẩm + quà tặng
      required: [id, orderId, productId, productName, quantity, unitPrice, isGift, createdAt]
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        productName: { type: string }
        unit: { type: string, nullable: true, description: "Đơn vị tính — sao chép từ Product tại thời điểm tạo đơn" }
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: number, minimum: 0 }
        discountPercent: { type: number, nullable: true, minimum: 0, maximum: 100, description: "Mới (Phase 3) — % Khuyến mại" }
        isGift: { type: boolean, default: false }
        createdAt: { type: string, format: date-time }
    CreateOrderItemInput:
      type: object
      required: [productId, productName, quantity, unitPrice]
      properties:
        productId: { type: string, format: uuid }
        productName: { type: string, minLength: 1 }
        unit: { type: string, description: "Đơn vị tính — sao chép từ Product đã chọn" }
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: number, minimum: 0 }
        discountPercent: { type: number, minimum: 0, maximum: 100, description: "Mới (Phase 3)" }
        isGift: { type: boolean, default: false }
...
    AccountProfile:
      type: object
      required: [id, username, isAdmin, createdAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        isAdmin: { type: boolean }
        employeeName:
          type: string
          nullable: true
          description: >
            Mới (Phase 3) — tên nhân viên liên kết (join ngược Employee.accountId=Account.id),
            phái sinh lúc trả response, KHÔNG lưu trên bảng Account. Null nếu tài khoản chưa được
            liên kết tới Employee nào (ví dụ các tài khoản module mặc định seed sẵn).
        createdAt: { type: string, format: date-time }
        permissions:
          type: array
          items: { $ref: "#/components/schemas/Permission" }
    CreateAccountInput:
      type: object
      required: [username]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 6, description: "Bỏ trống để dùng mật khẩu mặc định abc123" }
    UpdateAccountInput:
      type: object
      properties:
        username: { type: string, minLength: 1 }
        resetPassword: { type: boolean, description: "true = đặt lại mật khẩu về mặc định abc123" }
        newPassword: { type: string, minLength: 6, description: "Đặt mật khẩu tuỳ ý, ưu tiên hơn resetPassword nếu gửi cả hai" }
    Permission:
      type: object
      required: [module, canView, canEdit, canDelete]
      properties:
        module: { type: string }
        canView: { type: boolean }
        canEdit: { type: boolean }
        canDelete: { type: boolean }
    PermissionInput:
      type: object
      required: [module, canView, canEdit, canDelete]
      properties:
        module: { type: string }
        canView: { type: boolean }
        canEdit: { type: boolean }
        canDelete: { type: boolean }

    # ============================================================
    # Phase 2 — MỚI, theo MÔ TẢ NGHIỆP VỤ CỤ THỂ.pdf (V3.0)
    # ============================================================
...
    MarketingReport:
      type: object
      description: "Dữ liệu báo cáo (nhập tay định kỳ) — không phải nguồn sự thật, chỉ để đối chiếu (nguyên tắc #4)"
      required: [id, date, shift, product, market, team, adCost, messageCount, orderCount, revenue, revenueActual, createdAt]
      properties:
        id: { type: string, format: uuid }
        campaignId: { type: string, format: uuid, nullable: true }
        date: { type: string, format: date }
        shift: { type: string }
        product: { type: string }
        market: { type: string }
        team: { type: string }
        adCost: { type: number, description: "cpqc — chi phí quảng cáo" }
        messageCount: { type: integer }
        orderCount: { type: integer }
        revenue: { type: number }
        revenueActual: { type: number }
        warning: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateMarketingReportInput:
      type: object
      required: [date, shift, product, market, team, adCost, messageCount, orderCount, revenue, revenueActual]
      properties:
        campaignId: { type: string, format: uuid }
        date: { type: string, format: date }
        shift: { type: string, minLength: 1 }
        product: { type: string, minLength: 1 }
        market: { type: string, minLength: 1 }
        team: { type: string, minLength: 1 }
        adCost: { type: number, minimum: 0 }
        messageCount: { type: integer, minimum: 0 }
        orderCount: { type: integer, minimum: 0 }
        revenue: { type: number, minimum: 0 }
        revenueActual: { type: number, minimum: 0 }
        warning: { type: string }
    MarketingReportAuto:
      type: object
      description: "Mới — tính real-time từ Lead.sourcedBy + Order, không lưu bảng riêng"
      required: [employeeId, employeeName, leadCount, orderCount, revenue]
      properties:
        employeeId: { type: string, format: uuid }
        employeeName: { type: string }
        leadCount: { type: integer, description: "Số Lead nhân viên này tạo trong khoảng ngày" }
        orderCount: { type: integer, description: "Số đơn (đã xác nhận) từ các Lead của nhân viên này, phát sinh trong khoảng ngày" }
        revenue: { type: number, description: "Tổng doanh thu các đơn đó" }
    Customer:
      type: object
      required: [id, name, phone, customerType, blacklistStatus, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        phone: { type: string, description: "unique — chuẩn hoá khách hàng, tránh trùng lặp" }
        address: { type: string, nullable: true }
        city: { type: string, nullable: true }
        country: { type: string, nullable: true }
        customerType: { type: string, enum: [new, old] }
        blacklistStatus: { type: boolean, default: false }
        createdAt: { type: string, format: date-time }
    CreateCustomerInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        address: { type: string }
        city: { type: string }
        country: { type: string }
        customerType: { type: string, enum: [new, old], default: new }
    UpdateCustomerInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        address: { type: string }
        city: { type: string }
        country: { type: string }
        customerType: { type: string, enum: [new, old] }
        blacklistStatus: { type: boolean }
...
    Payment:
      type: object
      required: [id, orderId, amount, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        amount: { type: number }
        status: { type: string, enum: [pending, completed, failed] }
        method: { type: string, nullable: true, description: "vd MoneyGram, chuyển khoản" }
        feeBreakdown:
          type: object
          nullable: true
          additionalProperties: { type: number }
          description: "Object tự do cho các loại phụ phí (flight_fee, account_rental_fee...) — xem Quyết định thiết kế"
        reconciledAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    CreatePaymentInput:
      type: object
      required: [orderId, amount]
      properties:
        orderId: { type: string, format: uuid }
        amount: { type: number, minimum: 0 }
        method: { type: string }
    UpdatePaymentInput:
      type: object
      description: "status=completed yêu cầu đã có bill (enforce ở Service, không chỉ DTO — nguyên tắc #3)"
      properties:
        status: { type: string, enum: [pending, completed, failed] }
        feeBreakdown:
          type: object
          additionalProperties: { type: number }
        reconciledAt: { type: string, format: date-time }
...
    Employee:
      type: object
      description: "Hồ sơ nhân sự (nghiệp vụ) — khác Account (đăng nhập kỹ thuật), có thể liên kết qua accountId"
      required: [id, name, email, role, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        team: { type: string, nullable: true }
        branch: { type: string, nullable: true, description: "HN / HCM" }
        role: { type: string, enum: [MKT, Sale, CS, CSKH, KeToan, Admin] }
        status: { type: string, enum: [active, inactive] }
        accountId: { type: string, format: uuid, nullable: true, description: "Liên kết tới Account để đăng nhập, tuỳ chọn" }
        createdAt: { type: string, format: date-time }
    CreateEmployeeInput:
      type: object
      required: [name, email, role]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        team: { type: string }
        branch: { type: string }
        role: { type: string, enum: [MKT, Sale, CS, CSKH, KeToan, Admin] }
        accountId: { type: string, format: uuid }
    UpdateEmployeeInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        team: { type: string }
        branch: { type: string }
        role: { type: string, enum: [MKT, Sale, CS, CSKH, KeToan, Admin] }
        status: { type: string, enum: [active, inactive] }
        accountId: { type: string, format: uuid, nullable: true }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []